# feat: :sparkles: Allow markdown to include rhai script shortcodes.