This change processes the markdown with handlebars templates using scripts in the "shortcodes" directory, allowing content authors to include shortcodes in their markdown.

Thanks for considering this feature!