@bketelsen 2022-04-02
fixes #55