@bketelsen 2022-04-02
and a live version of this with tailwind styling instead of the bootstrap included in Bartholomew:
https://spin.brian.dev/