@bketelsen 2022-04-05
@technosophos @radu-matei @rajatjindal I made a small change to output the template rendering error if the shortcode script fails to parse. The whole PR could use some experienced Rust eyes to make sure I've not done something stupid.