@radu-matei 2022-04-05
I've been playing around with this PR and I'm seeing a non-trivial performance hit introduced (5ms to 15ms for rendering a blog entry):

```
# before this PR
❯ bombardier -c 1 -n 1000 http://localhost:3000/blog/2021-12-23-goals-of-bartholomew
Bombarding http://localhost:3000/blog/2021-12-23-goals-of-bartholomew with 1000 request(s) using 1 connection(s)
 1000 / 1000 100.00% 171/s 5s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec       173.89      34.11     257.64
  Latency        5.74ms     1.72ms    28.97ms
  HTTP codes:
    1xx - 0, 2xx - 1000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     1.51MB/s

# with this PR
❯ bombardier -c 1 -n 1000 http://localhost:3000/blog/2021-12-23-goals-of-bartholomew
Bombarding http://localhost:3000/blog/2021-12-23-goals-of-bartholomew with 1000 request(s) using 1 connection(s)
 1000 / 1000  100.00% 59/s 16s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec        60.34      20.44     118.85
  Latency       16.59ms     3.30ms    73.84ms
  HTTP codes:
    1xx - 0, 2xx - 1000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:   616.79KB/s
```

This is consistent even for pages that don't include shortcodes.
Currently exploring if there is something we can do to improve the rendering performance.