@vdice 2022-08-12
Now that https://github.com/fermyon/bartholomew/pull/91 is in, we have Rhai shortcode support.  It'll be included in the next release.  Thanks for coming up with the initial idea and PR @bketelsen!